# docs/openapi.yaml
...
openapi: "3.0.3"
info:
  title: Nuqta Backend API
  version: 1.0.0
  description: |
    REST JSON API for Nuqta POS / ERP system.
    All responses use the standard envelope:
    - Success: `{ ok: true, data: T }`
    - Error:   `{ ok: false, error: { code, message, details?, status? } }`
  contact:
    name: Nuqta Dev Team
servers:
  - url: /api/v1
    description: API v1 base path

# ─── Security Schemes ──────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT HS256 token.
        Access token TTL: 15 min (900s).
        Refresh token TTL: 7 days (604800s).
        Header: `Authorization: Bearer <accessToken>`

  # ─── Reusable Schemas ─────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - IDEMPOTENCY_REPLAY
            - INSUFFICIENT_STOCK
            - INVALID_STATE
            - INTERNAL_ERROR
        message:
          type: string
        details:
          description: Additional context (validation field errors, etc.)
        status:
          type: integer
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"

    # ── Pagination ──
...
    UserSafe:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        fullName: { type: string }
        phone: { type: string, nullable: true }
        role: { type: string, enum: [admin, cashier, manager, viewer] }
        isActive: { type: boolean }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
...
    Supplier:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        totalBalance: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
    CreateSupplierRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      additionalProperties: false
    UpdateSupplierRequest:
      type: object
      properties:
        name: { type: string }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        isActive: { type: boolean }

    # ── Sale ──
...
    LedgerEntry:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        amount: { type: integer }
        balance: { type: integer }
        description: { type: string }
        referenceId: { type: integer, nullable: true }
        referenceType: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    RecordPaymentRequest:
      type: object
      required: [amount, paymentMethod]
      properties:
        amount: { type: integer, minimum: 1 }
        paymentMethod: { type: string, enum: [cash, card, bank_transfer] }
        notes: { type: string }
        idempotencyKey: { type: string }
      additionalProperties: false
...
    CompanySettings:
      type: object
      properties:
        companyName: { type: string }
        companyAddress: { type: string, nullable: true }
        companyPhone: { type: string, nullable: true }
        companyEmail: { type: string, nullable: true }
        companyLogo: { type: string, nullable: true }
        taxNumber: { type: string, nullable: true }
...
    SetupWizardRequest:
      type: object
      properties:
        companyName: { type: string }
        currency: { type: string }
        modules: { type: object }
      additionalProperties: true

    # ── Barcode ──
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: "^\\d+$" }
      description: Numeric resource ID
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
    SearchParam:
      name: search
      in: query
      schema: { type: string }
    DateFromParam:
      name: dateFrom
      in: query
      schema: { type: string, format: date }
    DateToParam:
      name: dateTo
      in: query
      schema: { type: string, format: date }
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Idempotency key — required for financial write operations

  # ─── Reusable Responses ───────────────────────────────────────────
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Conflict (duplicate, state mismatch)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
paths:
...
  /suppliers:
    get:
      tags: [Suppliers]
      summary: List suppliers (paginated)
      operationId: listSuppliers
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: Supplier list
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Supplier" } }
                      total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Suppliers]
      summary: Create supplier
      operationId: createSupplier
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSupplierRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Supplier" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /suppliers/{id}:
    get:
      tags: [Suppliers]
      summary: Get supplier by ID
      operationId: getSupplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Supplier details
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Supplier" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Suppliers]
      summary: Update supplier
      operationId: updateSupplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSupplierRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Supplier" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Suppliers]
      summary: Delete supplier
      operationId: deleteSupplier
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: "null" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── SUPPLIER LEDGER ──────────────────────────────────────────────
  /supplier-ledger/{supplierId}:
    get:
      tags: [Supplier Ledger]
      summary: Get supplier ledger entries
      operationId: getSupplierLedger
      parameters:
        - name: supplierId
          in: path
          required: true
          schema: { type: integer }
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/LedgerEntry" } }
                      total: { type: integer }
                      balance: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /supplier-ledger/{supplierId}/payments:
    post:
      tags: [Supplier Ledger]
      summary: Record supplier payment
      operationId: recordSupplierPayment
      parameters:
        - name: supplierId
          in: path
          required: true
          schema: { type: integer }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecordPaymentRequest" }
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /supplier-ledger/reconcile:
    post:
      tags: [Supplier Ledger]
      summary: Reconcile supplier balances
      operationId: reconcileSupplierBalances
      parameters:
        - name: repair
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Reconciliation result
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ─── PURCHASES ────────────────────────────────────────────────────
...
  /settings/{key}:
    get:
      tags: [Settings]
      summary: Get setting by key
      operationId: getSetting
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Setting value
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: {}
    put:
      tags: [Settings]
      summary: Update setting by key
      operationId: updateSetting
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: "null" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /settings/company:
    get:
      tags: [Settings]
      summary: Get company settings
      operationId: getCompanySettings
      security: []
      responses:
        "200":
          description: Company settings
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/CompanySettings" }
    put:
      tags: [Settings]
      summary: Update company settings
      operationId: updateCompanySettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompanySettings" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/CompanySettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /settings/setup-wizard:
    post:
      tags: [Settings]
      summary: Complete setup wizard
      operationId: completeSetupWizard
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupWizardRequest" }
      responses:
        "200":
          description: Setup completed
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      completed: { type: boolean }

  # ─── USERS ────────────────────────────────────────────────────────
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: array, items: { $ref: "#/components/schemas/UserSafe" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, fullName, role]
              properties:
                username: { type: string, minLength: 1 }
                password: { type: string, minLength: 6 }
                fullName: { type: string, minLength: 1 }
                phone: { type: string, nullable: true }
                role: { type: string, enum: [admin, cashier, manager, viewer] }
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/UserSafe" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /users/{id}:
    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName: { type: string }
                phone: { type: string, nullable: true }
                role: { type: string, enum: [admin, cashier, manager, viewer] }
                isActive: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/UserSafe" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }