# src/routes/v1/suppliers/index.ts
import { FastifyPluginAsync } from "fastify";
import {
  GetSuppliersUseCase,
  GetSupplierByIdUseCase,
  CreateSupplierUseCase,
  UpdateSupplierUseCase,
  DeleteSupplierUseCase,
  type Supplier,
} from "@nuqta/core";
import {
  getSuppliersSchema,
  getSupplierByIdSchema,
  createSupplierSchema,
  updateSupplierSchema,
  deleteSupplierSchema,
} from "../../../schemas/suppliers.js";
import { requirePermission } from "../../../middleware/rbac.js";

const suppliers: FastifyPluginAsync = async (fastify) => {
  fastify.addHook("onRequest", fastify.authenticate);

  // GET /suppliers
  fastify.get(
    "/",
    {
      schema: getSuppliersSchema,
      preHandler: requirePermission("suppliers:read"),
    },
    async (request) => {
      const query = request.query as {
        search?: string;
        limit?: string;
        offset?: string;
      };
      const uc = new GetSuppliersUseCase(fastify.repos.supplier);
      const data = await uc.execute({
        search: query.search,
        limit: query.limit ? parseInt(query.limit, 10) : undefined,
        offset: query.offset ? parseInt(query.offset, 10) : undefined,
      });
      return { ok: true, data };
    },
  );

  // GET /suppliers/:id
  fastify.get<{ Params: { id: string } }>(
    "/:id",
    {
      schema: getSupplierByIdSchema,
      preHandler: requirePermission("suppliers:read"),
    },
    async (request) => {
      const id = parseInt(request.params.id, 10);
      const uc = new GetSupplierByIdUseCase(fastify.repos.supplier);
      const data = await uc.execute(id);
      return { ok: true, data };
    },
  );

  // POST /suppliers
  fastify.post(
    "/",
    {
      schema: createSupplierSchema,
      preHandler: requirePermission("suppliers:create"),
    },
    async (request) => {
      const body = request.body as Supplier;
      const uc = new CreateSupplierUseCase(fastify.repos.supplier);
      const data = await uc.execute(body);
      return { ok: true, data };
    },
  );

  // PUT /suppliers/:id
  fastify.put<{ Params: { id: string } }>(
    "/:id",
    {
      schema: updateSupplierSchema,
      preHandler: requirePermission("suppliers:update"),
    },
    async (request) => {
      const id = parseInt(request.params.id, 10);
      const body = request.body as Partial<Supplier>;
      const uc = new UpdateSupplierUseCase(fastify.repos.supplier);
      const data = await uc.execute(id, body);
      return { ok: true, data };
    },
  );

  // DELETE /suppliers/:id
  fastify.delete<{ Params: { id: string } }>(
    "/:id",
    {
      schema: deleteSupplierSchema,
      preHandler: requirePermission("suppliers:delete"),
    },
    async (request) => {
      const id = parseInt(request.params.id, 10);
      const uc = new DeleteSupplierUseCase(fastify.repos.supplier);
      await uc.execute(id);
      return { ok: true, data: null };
    },
  );
};

export default suppliers;